Move dbClient and migrations from dataservice to datamodel
