+++ Taskfile.yml
@@ -9,7 +9,6 @@ includes:
 tasks:
   dev:
     desc: Start the Docker Compose services in watch mode
-    deps: [build:bases]
     cmds:
       - docker compose --env-file ./workspace/config/.env up --watch --build 
 
@@ -30,6 +29,7 @@ tasks:
         platforms: [linux, darwin]
       - cmd: cmd /c copy .\\workspace\\config\\.env.template .\\workspace\\config\\.env
         platforms: [windows]
+      - task: build:bases
