+++ apps/servers/data_service/docker-compose.dataservice.yml
@@ -1,30 +1,32 @@
-graphcap_data_service:
-    container_name: graphcap_data_service
-    build:
-      context: ../../..
-      dockerfile: apps/servers/data_service/Dockerfile.data_service.dev
-    ports:
-      - "32550:32550"
-      - "32501:32501"  # Drizzle Kit Studio port
-    environment:
-      - NODE_ENV=${NODE_ENV:-development}
-      - PORT=32550
-      - DATABASE_URL=postgresql://${POSTGRES_USER}:${POSTGRES_PASSWORD}@${POSTGRES_HOST:-graphcap_postgres}:${POSTGRES_PORT:-5432}/${POSTGRES_DB}
-      - WORKSPACE_PATH=/workspace
-      - DEBUG=true  # Enable debug mode
-    volumes:
-      - ./workspace:/workspace
-      - ./apps/servers/data_service/src:/app/src  # Mount src directory directly
-    networks:
-      - graphcap
-    depends_on:
-      graphcap_postgres:
-        condition: service_healthy
-    healthcheck:
-      test: ["CMD", "wget", "--spider", "http://localhost:32550/health"]
-      interval: 5m
-      timeout: 10s
-      retries: 3
-      start_period: 30s
-    env_file:
-      - ./workspace/config/.env
+
+services:
+  graphcap_data_service:
+      container_name: graphcap_data_service
+      build:
+        context: ../../..
+        dockerfile: apps/servers/data_service/Dockerfile.dataservice.dev
+      ports:
+        - "32550:32550"
+        - "32501:32501"  # Drizzle Kit Studio port
+      environment:
+        - NODE_ENV=${NODE_ENV:-development}
+        - PORT=32550
+        - DATABASE_URL=postgresql://${POSTGRES_USER}:${POSTGRES_PASSWORD}@${POSTGRES_HOST:-graphcap_postgres}:${POSTGRES_PORT:-5432}/${POSTGRES_DB}
+        - WORKSPACE_PATH=/workspace
+        - DEBUG=true  # Enable debug mode
+      volumes:
+        - ../../../workspace:/workspace
+        - ./src:/app/apps/servers/data_service/src  # Mount src directory directly
+      networks:
+        - graphcap
+      depends_on:
+        graphcap_postgres:
+          condition: service_healthy
+      healthcheck:
+        test: ["CMD", "wget", "--spider", "http://localhost:32550/health"]
+        interval: 5m
+        timeout: 10s
+        retries: 3
+        start_period: 30s
+      env_file:
+        - ../../../workspace/config/.env
