--- apps/servers/data_service/drizzle.config.ts
@@ -1,20 +1,0 @@
-import { defineConfig } from 'drizzle-kit';
-import 'dotenv/config';
-
-export default defineConfig({
-  out: './src/db/migrations',
-  schema: './src/db/schema/index.ts',
-  dialect: 'postgresql',
-  dbCredentials: {
-    url: process.env.DATABASE_URL || '',
-  },
-  verbose: true,
-  strict: true,
-  tablesFilter: ['providers*', 'provider_*', 'batch_*'],
-  schemaFilter: 'public',
-  migrations: {
-    prefix: 'timestamp',
-    table: '__drizzle_migrations__',
-    schema: 'public'
-  }
-});
