+++ apps/servers/data_service/package.json
@@ -33,7 +33,9 @@
     "pino-pretty": "^13.0.0",
     "stoker": "1.4.2",
     "uuid": "^11.1.0",
-    "zod": "^3.24.2"
+    "zod": "^3.24.2",
+    "@graphcap/datamodel": "workspace:*",
+    "@graphcap/lib": "workspace:*"
   },
   "devDependencies": {
     "@iarna/toml": "^2.2.5",
