--- apps/servers/data_service/src/db/schema/index.ts
@@ -1,9 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-/**
- * Database schema exports
- * 
- * This file exports all database schema definitions for use with Drizzle ORM.
- */
-
-export * from '../../features/provider_config/db_providers';
-export * from './batch_queue';
