+++ packages/datamodel/entrypoint.sh
@@ -7,9 +7,18 @@ cd /app/packages/datamodel
 echo "Current working directory: $(pwd)"
 echo "Running database migrations..."
 # Directly run the drizzle-kit migrate command
-bunx drizzle-kit migrate --config=drizzle.config.ts
+# Use pnpm to ensure it uses the workspace version and config
+pnpm run db:migrate
 echo "Migrations complete."
 
-# Start Drizzle Studio, explicitly passing the config file
-echo "Starting Drizzle Studio..."
-bunx drizzle-kit studio --config=drizzle.config.ts --host 0.0.0.0 --port 59151 
+# Start Drizzle Studio only in development, explicitly passing the config file
+if [ "$NODE_ENV" = "development" ]; then
+    echo "Starting Drizzle Studio..."
+    # Use pnpm run to execute the script defined in package.json, passing host/port
+    pnpm run db:studio -- --host 0.0.0.0 --port 53151 &
+    echo "Drizzle Kit Studio started in background on 0.0.0.0:53151"
+else
+    echo "Not starting Drizzle Studio as NODE_ENV is not 'development'"
+fi
+
+echo "Datamodel entrypoint script finished."
