+++ packages/datamodel/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "graphcap-datamodel",
+	"name": "@graphcap/datamodel",
 	"version": "0.1.0",
 	"type": "module",
 	"main": "index.ts",
