+++ packages/lib/package.json
@@ -9,7 +9,9 @@
 		"@scalar/hono-api-reference": "^0.5.184",
 		"stoker": "1.4.2",
 		"pino": "^9.6.0",
-		"pino-pretty": "^13.0.0"
+		"pino-pretty": "^13.0.0",
+		"dotenv": "^16.4.7",
+		"dotenv-expand": "^12.0.1"
 	},
 	"devDependencies": {
 		"@types/pg": "^8.11.11",
